fix bugs to prevent overflow

+++ criu/cr-restore.c
@@ -2723,14 +2723,14 @@ static int restore_memory(pid_t pid)
 
 	timing_stop(TIME_MEMRST);
 	{
-		uint32_t time_us = timing_total(TIME_MEMRST);
+		uint64_t time_us = timing_total(TIME_MEMRST);
 		uint64_t total_size = (uint64_t)nr_restored_pages * PAGE_SIZE;
-		uint64_t throughput = 0;
+		double throughput = 0;
 
 		if (time_us > 0)
-			throughput = (total_size * 1000000) / (time_us * 1024 * 1024);
+			throughput = (double)(total_size * 1000000) / (time_us * 1024 * 1024);
 
-		pr_msg("Restore pages total time: %u us, size: %" PRIu64 " bytes, throughput: %" PRIu64 " MB/s\n",
+		pr_msg("Restore pages total time: %" PRIu64 " us, size: %" PRIu64 " bytes, throughput: %.2f MB/s\n",
 		       time_us, total_size, throughput);
 	}
